+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+- Unseal AreaTrigger and AreaTriggerUnityEvent
+
 ## [5.0.0] - 2026-07-31
 ### Added
 - AreaTriggerUnityEvent component